+++ synthex/jobs_api.py
@@ -1,9 +1,8 @@
 from .api_client import APIClient
-from typing import Any, List, Optional
+from typing import Any, List
 import csv
 import os
 import time
-import logging
 
 from .models import ListJobsResponseModel, JobOutputType, JobOutputSchemaDefinition, ActionResult, \
     JobStatusResponseModel, SuccessResponse
@@ -14,9 +13,6 @@ from .config import config
 from .endpoints import GET_JOB_DATA_ENDPOINT, GET_JOB_STATUS_ENDPOINT
 
 
-logger = logging.getLogger(__name__)
-
-
 @auto_validate_methods
 class JobsAPI:
     """
@@ -25,8 +21,6 @@ class JobsAPI:
     
     def __init__(self, client: APIClient):
         self._client: APIClient = client
-        # TODO: add support for multiple concurrent jobs; this should require launching each job in a separate thread.
-        self._current_job_id: Optional[str] = None
         
     def list(self, limit: int = 10, offset: int = 0) -> ListJobsResponseModel:
         """
@@ -107,9 +101,8 @@ class JobsAPI:
         
         response = self._client.post(f"{CREATE_JOB_WITH_SAMPLES_ENDPOINT}", data=data)
         
-        # Store id of the current job.
-        self._current_job_id = response.data
-        
+        print(f"Started job with ID {response.data}")
+
         if response.data is None:
             raise ValidationError("Response data is None, expected a valid job ID.")
         return response.data
@@ -144,7 +137,6 @@ class JobsAPI:
         # Keep fetching data as long as the status code is 206, meaning that there is more data;
         # When the status code changes to 200, that means there is no more data to fetch.
         keep_polling = True
-        # TODO: a timeout is needed here: if a job has not received any data within x seconds, error out
         while keep_polling:
             # Poll every JOB_DATA_POLLING_INTERVAL seconds
             time.sleep(config.JOB_DATA_POLLING_INTERVAL)
@@ -168,9 +160,9 @@ class JobsAPI:
                     # Append each dict as a row.
                     writer.writerows(data)
 
-                    logger.info(f"{len(data)} datapoints written for job {job_id}")
+                    print(f"{len(data)} datapoints written for job {job_id}")
 
-        logger.info(f"All datapoints for job {job_id} have been written")
+        print(f"All datapoints for job {job_id} have been written")
 
         return True
 
@@ -219,20 +211,18 @@ class JobsAPI:
 
         return ActionResult(
             success=True,
-            message=f"Job started successfully. Output will be saved to '{output_path}' upon completion.",
+            message=f"Job with ID {job_id} was started successfully. Output will be saved to '{output_path}'.",
         )
 
-    def status(self) -> JobStatusResponseModel:
+    def status(self, job_id: str) -> JobStatusResponseModel:
         """
-        Check the status of the job that is currently running.
+        Check the status of a job.
+        Args:
+            job_id (str): The ID of the job to check the status of.
         Returns:
             JobStatusResponseModel: A model containing the status of the job.
         """
-        
-        # No job has been started yet.
-        if not self._current_job_id:
-            raise ValidationError("No job is currently running.")
-                
-        response = self._client.get(GET_JOB_STATUS_ENDPOINT(self._current_job_id))
-                
+
+        response = self._client.get(GET_JOB_STATUS_ENDPOINT(job_id))
+
         return JobStatusResponseModel.model_validate(response.data)