+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Release v0.5.0 - September 4, 2025
+
+### Modified
+
+- Replaced logger calls with print statements in the `_get_data_and_write_to_file` function.
+- Updated the `status` method in `JobsAPI` to accept a `job_id` parameter, allowing users to check the status of any job by its ID instead of only the currently running job.
+
 ## Release v0.4.0 - September 3, 2025
 
-Firt release.
+First release.